test(conformance): shared counter-storm harness + go_go parity; document mixed-storm Go double-apply (#1043)
Refactor the same-doc counter STORM into a single cluster-agnostic driver
(`support::run_counter_storm`) shared by the Rust regression guards and the
Go-parity twins, drop the redundant `nodes` arg (derived from `node_deltas`),
and add `parity_counter_storm_go_go` as the Go-only oracle (asserts, wired into
the blocking go-compat CI leg).

The mixed Rust<->Go storm (`parity_counter_storm_mixed`) surfaced an
intermittent over-count. Investigation (link-level DAG dumps + an instrumented
Go binary) established it is a timing-sensitive DOUBLE-APPLY on the GO reference
side, not a Rust value bug: all three nodes hold a byte-identical DAG, Rust
materializes the exact sum, and Go peers materialize +k because Go's
`coreblock.ProcessBlock` runs `incrementValue` (RMW) unconditionally with no
per-block IsMerged guard (dedup is purely structural via loadComposites),
whereas Rust's counter merge guards on is_merged(cid). It never reproduces in
pure go<->go (0/15) and is suppressed by Go-side instrumentation (Heisenbug).

The mixed test stays asserting (encodes the correct expectation) but is
#[ignore]'d and EXCLUDED from the blocking go-compat CI filter (now selects
parity_counter_storm_go_go) until the race is resolved. Full root cause and a
candidate Rust-side mitigation (match Go's concurrent dual-topic publish) are in
#1043.
